Ajout de commentaire et de facilité d'utilisation
