+++ src/climatonix/API/ParamaterStringBuilder.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package climatonix;
+package climatonix.API;
 
 import java.io.UnsupportedEncodingException;
 import java.net.URLEncoder;
@@ -16,8 +16,10 @@ import java.util.Map;
 public class ParamaterStringBuilder {
     public static String getParamsString(Map<String, String> params) 
       throws UnsupportedEncodingException{
+        // Le résultat
         StringBuilder result = new StringBuilder();
  
+        // Transformer le Map dans le format nécessaire pour le URL
         for (Map.Entry<String, String> entry : params.entrySet()) {
           result.append(URLEncoder.encode(entry.getKey(), "UTF-8"));
           result.append("=");
@@ -25,7 +27,10 @@ public class ParamaterStringBuilder {
           result.append("&");
         }
  
+        // Convertir dans un String
         String resultString = result.toString();
+        
+        // Retourner le String
         return resultString.length() > 0
           ? resultString.substring(0, resultString.length() - 1)
           : resultString;
