+++ src/climatonix/Climatonix.java
@@ -19,20 +19,9 @@ public class Climatonix {
     public static void main(String[] args) {
         // TODO code application logic here
         
-        JSONObject jSONObject;
         
-        
-        try {
-            jSONObject = APIUtils.request("weather");
-        } catch(Exception e) {
-            e.printStackTrace();
-        }
-        
-        int h = 0;
     }
     
-    public double kelvinToCelcius(double kelvin) {
-        return kelvin - 273.15;
-    }
+    
     
 }
